refactor: move contest utils to src/lib/contests/utils/
